third_party/WebKit/LayoutTests/dom/attr/set-attr-value-no-DOMSubtreeModified.html

<!DOCTYPE html>
<script src="../../resources/js-test.js"></script>
<div id="testDiv"></div>
<script>
    description("Tests that setting Attr.value does not fire a DOMSubtreeModified event.");

    var testDiv = document.getElementById("testDiv");
    var wasDOMSubtreeModifiedFired = false;
    testDiv.attributes[0].addEventListener("DOMSubtreeModified", function() {
        wasDOMSubtreeModifiedFired = true;
    }, false);

    shouldBeFalse("wasDOMSubtreeModifiedFired");
    testDiv.attributes[0].value = "testDiv2";
    shouldBeFalse("wasDOMSubtreeModifiedFired");
</script>
