third_party/WebKit/LayoutTests/dom/attr/set-attr-value-no-DOMSubtreeModified-expected.txt

Tests that setting Attr.value does not fire a DOMSubtreeModified event.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS wasDOMSubtreeModifiedFired is false
PASS wasDOMSubtreeModifiedFired is false
PASS successfullyParsed is true

TEST COMPLETE
